Update Lineweaver Burke.R
In this version, I changed the theme for the graph from theme_linedraw() (black grid that is visually overwhelming) to theme_bw() (which features a grey grid that doesn't stand out as much so that data points and line can be seen more clearly). 

I also changed the position of the text featuring the equation of the line of best fit. If you specify its position through set x and y coordinates, this will mean it could end up all over the place depending on the size of the values in your data. 

Therefore, I set the text to be at a fixed position that will always be at 1/2 of the maximum x and y values, and slightly justified to the right.
Due to the nature of Lineweaver-Burk* and Hanes-Woolf plots, this should mean that the text should never cross the line and remain legible...

*according to wikipedia his name doesn't have an "e" at the end.